refactor: MemberForAuth 클래스를 Member 클래스와 통합
