+++ src/main/java/nextstep/member/domain/Member.java
@@ -1,16 +1,46 @@
 package nextstep.member.domain;
 
+import lombok.AllArgsConstructor;
+import lombok.Builder;
 import lombok.Getter;
+import org.springframework.security.crypto.bcrypt.BCryptPasswordEncoder;
+import org.springframework.security.crypto.password.PasswordEncoder;
 
 @Getter
-public class Member extends MemberForAuth {
+@Builder
+@AllArgsConstructor
+public class Member {
 
-    protected final String name;
-    protected final String phone;
+    private final Long id;
+    private final String name;
+    private final String phone;
+    private final String username;
+    private String password;
 
-    public Member(String username, String password, String name, String phone) {
-        super(username, password);
-        this.name = name;
-        this.phone = phone;
+    public final boolean matchPassword(Member target) {
+
+        return PasswordUtil.matches(target.getPassword(), this.password);
+    }
+
+    public final void encryptPassword() {
+        this.password = PasswordUtil.encrypt(this.password);
+    }
+
+    private static final class PasswordUtil {
+
+        private static final PasswordEncoder passwordEncoder = new BCryptPasswordEncoder();
+
+        private PasswordUtil() {
+        }
+
+        public static boolean matches(String rawPassword, String encryptedPassword) {
+
+            return passwordEncoder.matches(rawPassword, encryptedPassword);
+        }
+
+        public static String encrypt(String password) {
+
+            return passwordEncoder.encode(password);
+        }
     }
 }
